added conditions helper functions
